fix for cursor timeout within docker swarm deployment
docker swarm uses VIP and router mesh even for global service. I.e.,
engine may load-balance requests to db to a different node even if
there is a resource available on the local node. As a result, the socket
where the cursor is acquired may be on one mongos instance and then the
next request on the cursor may be sent to another mongos instance
where cursor is unknown.
Fix is to move db from global service to unique service running in dnsrr
and pinned to each node. pymongo client should be able to load-balance
between separate mongos instances if hostnames are unique